update group 에서도 사용가능하게 변경
